Extract ingester_test utility methods and structs in a separate source
